test: Fixed AT 1.4, 1.5, 1.6
This also includes the appropriate id changes in the front end to allow for all of the options to be obtainable
